Add ability to keep profile after execution

+++ anki_testing.py
@@ -10,8 +10,8 @@ from warnings import warn
 
 
 @contextmanager
-def _temporary_user(dir_name: str, name: str="__Temporary Test User__",
-                    lang: str="en_US"):
+def _temporary_user(dir_name: str, name: str, lang: str, keep: bool):
+
     from aqt.profiles import ProfileManager
 
     # prevent popping up language selection dialog
@@ -26,7 +26,7 @@ def _temporary_user(dir_name: str, name: str="__Temporary Test User__",
 
     pm.setupMeta()
 
-    if name in pm.profiles():
+    if not keep and name in pm.profiles():
         warn(f"Temporary user named {name} already exists")
     else:
         pm.create(name)
@@ -35,22 +35,26 @@ def _temporary_user(dir_name: str, name: str="__Temporary Test User__",
 
     yield name
 
-    pm.remove(name)
+    if not keep:
+        pm.remove(name)
+
     ProfileManager._setDefaultLang = original
 
 
 @contextmanager
-def _temporary_dir(tmp_path: str, name: str):
+def _temporary_dir(tmp_path: str, name: str, keep: bool):
     path = os.path.join(tmp_path, name)
     yield path
-    shutil.rmtree(path)
+    if not keep:
+        shutil.rmtree(path)
 
 
 @contextmanager
 def anki_running(anki_path: str="anki_root",
                  tmp_path: str=tempfile.gettempdir(),
                  lang: str="en_US",
-                 keep_profile):
+                 profile_name="__Temporary Test User__",
+                 keep_profile=False):
 
     if anki_path and anki_path not in sys.path:
         sys.path.insert(0, anki_path)
@@ -59,8 +63,8 @@ def anki_running(anki_path: str="anki_root",
     from aqt import _run
 
     # we need a new user for the test
-    with _temporary_dir(tmp_path, "anki_temp_base") as dir_name:
-        with _temporary_user(dir_name, lang=lang) as user_name:
+    with _temporary_dir(tmp_path, "anki_temp_base", keep_profile) as dir_name:
+        with _temporary_user(dir_name, lang, profile_name, keep_profile) as user_name:
             app = _run(argv=["anki", "-p", user_name,
                              "-b", dir_name], exec=False)
             yield app
